@@ -26,9 +26,9 @@ class TestNeighbourhoodBlock(TestData):
 
             def with_nulls(vals):
                 vals = vals.copy()
-                vals.iloc[
-                    np.random.choice(len(df), size=nan_count, replace=False)
-                ] = np.nan
+                vals.iloc[np.random.choice(len(df), size=nan_count, replace=False)] = (
+                    np.nan
+                )
                 return vals
 
             return df.copy() if nan_count <= 0 else df.apply(with_nulls)
